Attempting to fix ci with path renaming.
