+++ game/lib/home_page/home_page_widget.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:sliding_up_panel/sliding_up_panel.dart';
 import 'package:velocity_x/velocity_x.dart';
-import 'package:game/widgets/nav_bar.dart';
-import 'package:game/widgets/nav_btn.dart';
+import 'package:game/widget/nav_bar.dart';
+import 'package:game/widget/nav_btn.dart';
 
 class HomePageWidget extends StatefulWidget {
   HomePageWidget({Key? key}) : super(key: key);
