+++ game/lib/leaderboard/leaderboard_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:game/widgets/back_btn.dart';
-import 'package:game/widgets/leaderboard_cell.dart';
+import 'package:game/widget/back_btn.dart';
+import 'package:game/widget/leaderboard_cell.dart';
 
 class LeaderboardWidget extends StatefulWidget {
   LeaderboardWidget({Key? key}) : super(key: key);
