+++ game/lib/visited_places/visited_places_widget.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:game/widgets/back_btn.dart';
+import 'package:game/widget/back_btn.dart';
 
 class VisitedPlacesWidget extends StatefulWidget {
   VisitedPlacesWidget({Key? key}) : super(key: key);
